+++ src/command_modules/azure-cli-acs/azure/cli/command_modules/acs/_validators.py
@@ -90,3 +90,17 @@ def validate_linux_host_name(namespace):
     if not found:
         raise CLIError('--name cannot exceed 63 characters and can only contain '
                        'letters, numbers, or dashes (-).')
+
+
+def validate_connector_name(namespace):
+    """Validates a string as a legal connector name.
+
+    This validation will also occur server-side in the kubernetes, but that may take
+    for a while. So it's more user-friendly to validate in the CLI pre-flight.
+    """
+    # https://github.com/kubernetes/community/blob/master/contributors/design-proposals/architecture/identifiers.md
+    regex = re.compile(r'^[a-z0-9]([a-z0-9\-]*[a-z0-9])?$')
+    found = regex.findall(namespace.connector_name)
+    if not found:
+        raise CLIError('--connector-name must consist of lower case alphanumeric characters or dashes (-), '
+                       'and must start and end with an alphanumeric character.')